requirements.txt
asgiref==3.6.0
autopep8==1.7.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.2.0
click==8.1.3
cloudinary==1.34.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1.7
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-js-asset==2.1.0
filelock==3.9.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pipenv==2023.2.18
platformdirs==3.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.6
pycodestyle==2.9.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.6.2
urllib3==1.26.16
virtualenv==20.19.0
virtualenv-clone==0.5.7
Werkzeug==2.3.4
whitenoise==6.5.0
